Moved dashboard fetch to the actual dashboard
