+++ src/layouts/dashboard.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, createContext } from "react";
+import React, { useState, createContext } from "react";
 import { ThemeProvider } from '@emotion/react';
 import { styled } from '@mui/material/styles';
 import { theme } from "assets/styles/AppTheme.js";
@@ -15,15 +15,7 @@ import D3Chart from "../components/D3Container/D3Chart.js";
 import Welcome from "../components/Cards/CardWelcome.js";
 import Stars from "../components/Cards/CardStars.js";
 
-import { dataList } from '../components/D3Container/DemoData';
-
 export const datastoreContext = createContext("");
-const axios = require('axios').default;
-
-const instance = axios.create({
-  baseURL: `${process.env.REACT_APP_HEDIS_MEASURE_API_URL}`,
-  timeout: 1000
-});
 
 const Item = styled(Paper)(({ theme }) => ({
   ...theme.typography.body2,
@@ -34,58 +26,6 @@ const Item = styled(Paper)(({ theme }) => ({
 }));
 
 export default function Admin() {
-  const [measures, setMeasures] = useState([]);
-  let valueArray = [];
-
-  useEffect(() => {
-    instance.get('measures/search', {
-      params: {
-        //memberId: '6dccff7c-db25-a27b-d718-7189b766b218',
-        measurementType: 'drre'
-      }
-    })
-    .then(res => {
-      if (res && res.data.length > 30) {
-        let numeratorValues = [];
-        let denominatorValues = [];
-        for (const patientJSON in res.data) {
-          const patient = JSON.parse(patientJSON);
-          for (const patientField in patient) {
-            if (patientField.startsWith('Numerator')) {
-              let numCount = 0;
-              if (patientField !== 'Numerator') {
-                numCount = patientField.replace('Numerator ', '') - 1
-              }
-              numeratorValues[numCount] += patient[patientField]
-            }
-            else if (patientField.startsWith('Denominator')) {
-              let denCount = 0;
-              if (patientField !== 'Denominator') {
-                denCount = patientField.replace('Denominator ', '') - 1
-              }
-              denominatorValues[denCount] += patient[patientField]
-            }
-          }
-        }
-
-        const measureSize = numeratorValues.length();
-        const currentDate = new Date();
-        const dateString = currentDate.getFullYear() + '-' + (currentDate.getMonth() + 1) + '-' + currentDate.getDate()
-        for (let i = 0; i < measureSize; i++) {
-          const numerator = numeratorValues[i];
-          const denominator = denominatorValues[i];
-          valueArray.push({
-            name: 'drre ' + (i + 1),
-            date: dateString,
-            value: (numerator/denominator) * 100
-          });
-        }
-      }
-    })
-    .catch(function (error) {
-      console.log(error);
-    })
-  }, []);
 
   const [datastore, setDatastore] = useState("");
 
@@ -114,7 +54,7 @@ export default function Admin() {
                 </Grid>
                 <Grid item xs={12}>
                   <Item>
-                    <D3Chart measures={dataList} />
+                    <D3Chart/>
                   </Item>
                 </Grid>
                 <Grid item xs={12}>
